Issue #37: Lorax would fail if old artificats existed
This fixes issues when rebuilding without clearing out tmp/. We still
need a decision on when to clear out the build directory. For example,
the x86_64.log will be appended to after each build. Also, some LiveOS
directories remain after mksquashfs is called. This will be addressed in
issue #79.

+++ packages/lorax/lorax-19.6.45/src/pylorax/sysutils.py
@@ -96,7 +96,8 @@ def cpfile(src, dst):
 def mvfile(src, dst):
     if os.path.isdir(dst):
         dst = joinpaths(dst, os.path.basename(src))
-    os.rename(src, dst)
+    if not os.path.exists(dst):
+        os.rename(src, dst)
     return dst
 
 def remove(target):
@@ -106,4 +107,4 @@ def remove(target):
         os.unlink(target)
 
 def linktree(src, dst):
-    runcmd(["/bin/cp", "-alx", src, dst])
+    runcmd(["/bin/cp", "-alxf", src, dst])
